refactor(ProjectionThread): split? ipc logic to IPCConnection
