+++ ProjectionThread.hpp
@@ -12,18 +12,19 @@
 
 #include "UnixSocket.hpp"
 
+#include "IPCConnection.hpp"
 #include "messages/projector.h"
 
 #include "ulalaca.hpp"
 
-using MallocFreeDeleter = std::function<void(void *)>;
 
 class ProjectionThread {
 public:
     explicit ProjectionThread(
         XrdpUlalaca &xrdpUlalaca,
-        UnixSocket &socket
+        const std::string &socketPath
     );
+    ProjectionThread(ProjectionThread &) = delete;
     
     void start();
     void stop();
@@ -31,56 +32,12 @@ public:
     void handleEvent(XrdpEvent &event);
 private:
     void mainLoop();
-    void ioLoop();
-    
-    std::unique_ptr<ProjectorMessageHeader, MallocFreeDeleter> nextHeader();
-    
-    template<typename T>
-    std::unique_ptr<T, MallocFreeDeleter> read(size_t size) {
-        assert(size != 0);
-       
-        auto promise = std::promise<std::unique_ptr<uint8_t>>();
-        {
-            std::scoped_lock<std::mutex> scopedReadTasksLock(_readTasksLock);
-            _readTasks.emplace(size, promise);
-        }
-        auto pointer = promise.get_future().get();
-    
-        return std::move(std::unique_ptr<T, MallocFreeDeleter>(
-            reinterpret_cast<T *>(pointer.release()),
-            free
-        ));
-    }
-    
-    void write(const void *pointer, size_t size);
-    
-    template <typename T>
-    void writeMessage(uint16_t messageType, T message) {
-        auto header = ProjectorMessageHeader {
-            (uint16_t) messageType,
-            0, 0,
-            0,
-            sizeof(T)
-        };
-        
-        write(&header, sizeof(header));
-        write(&message, sizeof(T));
-    }
-    
     
     XrdpUlalaca &_xrdpUlalaca;
-    UnixSocket &_socket;
+    IPCConnection _ipcConnection;
     
     bool _isTerminated;
     std::thread _projectorThread;
-    std::thread _ioThread;
-    
-    
-    std::mutex _writeTasksLock;
-    std::mutex _readTasksLock;
-    
-    std::queue<std::pair<size_t, std::unique_ptr<uint8_t, MallocFreeDeleter>>> _writeTasks;
-    std::queue<std::pair<size_t, std::promise<std::unique_ptr<uint8_t>> &>> _readTasks;
 };
 
 #endif //ULALACA_PROJECTIONTHREAD_HPP
